Full working demo of downloading new packages and adding them to commits
A branch holds a package list.  When a client requests a file that
doesn't exist in the branch, it first checks the master cache (shared
by all branches) and if it isn't there, it fetches it from CPAN. Then
it adds the file to the master cache, and adds the file ot the current
branch.  It delays the commit of these changes by some number of
seconds, 10 by default.