New design for mirror files
I now have the package list uncompressed in each branch, and
compress it when served.  Then, each author .tar.gz can either be
added as a file, or added as an extracted directory which will be
re-assembled into a .tar.gz on request.

The controller is updated, and works with

    cpanm --from=http:127.0.0.1:3000/main [Module::Name]

Not well tested yet, though.