Began writing the cpangit-add command
One complication I immediately ran into is that I want to run this
on the working directory a lot during testing, but later I will want
to run this from the Mojolicious web app where there is no working
directory.  I'm now planning to have two different modes of operation
for the objects that represent the cpan file tree, where it can either
commit directly to the branch, or else write files and add them to the
index without committing.

I'm also planning to make the class representing the cache of dist files
completely separate from the class that represents a CPAN mirror tree.