Refactoring so that I can easily alter trees
If a MirrorTree refers to the branch of the working directory, I want changes
to that tree to be applied to the filesystem and index.  Otherwise they need
cached in memory and maybe serialized until the user is ready to commit them
and then a new tree needs written out.