Finished refactor of ::MutableTree, ::MirrorTree
MutableTree is generically able to handle reading and writng a
branch, a tree detached from a branch, or the current index.
I was really hoping to generically use an Index object, but the
Git::Raw implementation doesn't allow me to add anything to an
index unless it is the official index of the repo.  I did find a
hacky workaround where I could take an independent index, swap it
with the repo's index, add content, then swap it back, but that
was ugly enough I fell back to my previous design where I hold a
tree of hashrefs and then use TreeBuilder to assemble them before
a commit.

I then updated MirrorTree to parse and write the packages list
using these methods.

The DistCache is not updated yet.