Redesign MirrorTree a bit; working test case of mirroring from CPAN
My new plan is that MirrorTree will be used both as the DarkPAN file
tree used by Dockerfiles, and also as the cache of a remote mirror
such as www.clan.org.  The new attribute upstream_url controls whether
it should pull things down on demand, or require the user to
intentionally add the packages.

I also now have a working test case of downloading the package list
and the package Crypt::DES from www.cpan.org.  (Crypt::DES is the
problem package that needs patched and initially triggered this effort)