Implemented process_distfile
Processing CPAN dist files is sort of hard.  I don't need a 100%
solution here, just enough to help us fetch dependent modules.
I am extracting the tar file to get the META.json (or META.yml)
and then if the module doesn't declare its 'provides' list, I use
Module::Metadata to build a best guess.

I also added an un-tested option to extract the tar file directly
into the Git repo.  However, then I discovered that some CPAN dists
can be .zip files, and a very few are bar ".pm.gz" files.  I'm not
going to worry about those for now.

I notice that the handling of the Git repo is broken if modifying
the branch which is checked out.  I have logic to handle it, but it
seems not to work right.