Tests and bugfixes for ::ArchiveTree
I fixed bugs in version parsing, and improved the logic of
combine_version_requirements to combine more variations.

I gave MutableTree the ability to automatically look up the branch
by name, and automatically grab the tree from the branch.

Fixed a bug in the 'package_details' attribute, and tested that it
parses correctly.