ArchiveTree can successfully import package requirements
It turned out there's a core module CPAN::Meta::Requirements that
already does a lot fo what I was doing with version comparisons.
I updated all my code to use that, where possible.  Now the 'reqs'
parameter to `import_modules` can be that object or coerces a hash
into an object.

Then I worked through all the logic of filtering the requirements
by the modules already present in the ArchiveTree or in the perl
CoreList, and added logging that is fairly precise to only what is
getting added.  I also added logging for the 'recommends' lists.

I also cleaned up the InGit interface of `get_archive_tree` and
`create_archive_tree`, and added a special case to `lookup_tree`
for getting the 'HEAD' branch and setting `use_workdir` on the tree
returned by `get_archive_tree`.