Re-implement the CPAN server as a Mojolicious::Controller
This keeps more of the logic in .pm libs which can be subclassed.
The `mount` class method lets a Mojo app easily add it to a
sub-path.

I changed it to only serve the specific paths known to be public,
like th epackages.details and the author files.  This protects the
config file at `/cpan_ingit.json` in case it contains secrets.